+++ app/controllers/digitalEnvelope.controller.js
@@ -82,6 +82,34 @@ exports.findOne = (req, res) => {
     });
 };
 
+exports.findByInvitationId = (req, res) => {
+  const id = req.params.id;
+
+  DigitalEnvelope.findAll({
+    where: { invitationId: id },
+    include: { model: db.invitation },
+    order: [["updatedAt", "DESC"]],
+    attributes: { exclude: ["createdAt", "updatedAt"] },
+  })
+    .then((data) => {
+      if (data) {
+        res.status(200).send({
+          message: "success",
+          data: data,
+        });
+      } else {
+        res.status(404).send({
+          message: `Cannot find DigitalEnvelope with id=${id}.`,
+        });
+      }
+    })
+    .catch((err) => {
+      res.status(500).send({
+        message: "Error retrieving DigitalEnvelope with id=" + id,
+      });
+    });
+};
+
 // Update a DigitalEnvelope by the id in the request
 exports.update = (req, res) => {
   const id = req.params.id;
@@ -90,7 +118,7 @@ exports.update = (req, res) => {
     attributes: { exclude: ["createdAt", "updatedAt"] },
   })
     .then((data) => {
-      if (req.file !== undefined) {
+      if (req.file !== undefined && data.image !== "") {
         fs.unlink("./upload/images/" + data.image, (err) => {
           if (err) throw err;
         });
@@ -98,7 +126,9 @@ exports.update = (req, res) => {
       data
         .update({
           name: req.body.name,
-          image: req.file === undefined ? data.image : req.file.filename,
+          number: req.body.number,
+          published: req.body.published,
+          image: req.file === undefined ? "" : req.file.filename,
         })
         .then(() => {
           res.status(200).send({
