+++ app/controllers/music.controller.js
@@ -84,7 +84,7 @@ exports.update = (req, res) => {
     attributes: { exclude: ["createdAt", "updatedAt"] },
   })
     .then((data) => {
-      if (req.file !== undefined) {
+      if (req.file !== undefined && data.image !== "") {
         fs.unlink("./upload/files/songs/" + data.song, (err) => {
           if (err) throw err;
         });
